Create rename.py
TESTED.
This program renames simulation files. If you output every Xth run of your simulation, i.e. the snapshot numbers follow a sequence with a step X (start, start+X, start+2X, ...), with this program you can change the ordering to be an array of integers with a step X=1.